+++ src/quantum/azext_quantum/vendored_sdks/azure_mgmt_quantum/models/__init__.py
@@ -6,103 +6,78 @@
 # Changes may cause incorrect behavior and will be lost if the code is regenerated.
 # --------------------------------------------------------------------------
 
-try:
-    from ._models_py3 import CheckNameAvailabilityParameters
-    from ._models_py3 import CheckNameAvailabilityResult
-    from ._models_py3 import ErrorAdditionalInfo
-    from ._models_py3 import ErrorDetail
-    from ._models_py3 import ErrorResponse
-    from ._models_py3 import OfferingsListResult
-    from ._models_py3 import Operation
-    from ._models_py3 import OperationDisplay
-    from ._models_py3 import OperationsList
-    from ._models_py3 import PricingDetail
-    from ._models_py3 import PricingDimension
-    from ._models_py3 import Provider
-    from ._models_py3 import ProviderDescription
-    from ._models_py3 import ProviderProperties
-    from ._models_py3 import ProviderPropertiesAad
-    from ._models_py3 import ProviderPropertiesManagedApplication
-    from ._models_py3 import QuantumWorkspace
-    from ._models_py3 import QuantumWorkspaceIdentity
-    from ._models_py3 import QuotaDimension
-    from ._models_py3 import Resource
-    from ._models_py3 import SkuDescription
-    from ._models_py3 import SystemData
-    from ._models_py3 import TagsObject
-    from ._models_py3 import TargetDescription
-    from ._models_py3 import TrackedResource
-    from ._models_py3 import WorkspaceListResult
-except (SyntaxError, ImportError):
-    from ._models import CheckNameAvailabilityParameters  # type: ignore
-    from ._models import CheckNameAvailabilityResult  # type: ignore
-    from ._models import ErrorAdditionalInfo  # type: ignore
-    from ._models import ErrorDetail  # type: ignore
-    from ._models import ErrorResponse  # type: ignore
-    from ._models import OfferingsListResult  # type: ignore
-    from ._models import Operation  # type: ignore
-    from ._models import OperationDisplay  # type: ignore
-    from ._models import OperationsList  # type: ignore
-    from ._models import PricingDetail  # type: ignore
-    from ._models import PricingDimension  # type: ignore
-    from ._models import Provider  # type: ignore
-    from ._models import ProviderDescription  # type: ignore
-    from ._models import ProviderProperties  # type: ignore
-    from ._models import ProviderPropertiesAad  # type: ignore
-    from ._models import ProviderPropertiesManagedApplication  # type: ignore
-    from ._models import QuantumWorkspace  # type: ignore
-    from ._models import QuantumWorkspaceIdentity  # type: ignore
-    from ._models import QuotaDimension  # type: ignore
-    from ._models import Resource  # type: ignore
-    from ._models import SkuDescription  # type: ignore
-    from ._models import SystemData  # type: ignore
-    from ._models import TagsObject  # type: ignore
-    from ._models import TargetDescription  # type: ignore
-    from ._models import TrackedResource  # type: ignore
-    from ._models import WorkspaceListResult  # type: ignore
+from ._models import APIKeys
+from ._models import ApiKey
+from ._models import CheckNameAvailabilityParameters
+from ._models import CheckNameAvailabilityResult
+from ._models import ErrorAdditionalInfo
+from ._models import ErrorDetail
+from ._models import ErrorResponse
+from ._models import ListKeysResult
+from ._models import Operation
+from ._models import OperationDisplay
+from ._models import PricingDetail
+from ._models import PricingDimension
+from ._models import Provider
+from ._models import ProviderDescription
+from ._models import ProviderProperties
+from ._models import ProviderPropertiesAad
+from ._models import ProviderPropertiesManagedApplication
+from ._models import QuantumWorkspace
+from ._models import QuantumWorkspaceIdentity
+from ._models import QuotaDimension
+from ._models import Resource
+from ._models import SkuDescription
+from ._models import SystemData
+from ._models import TagsObject
+from ._models import TargetDescription
+from ._models import TrackedResource
+from ._models import WorkspaceResourceProperties
 
-from ._azure_quantum_management_client_enums import (
-    CreatedByType,
-    ProvisioningStatus,
-    ResourceIdentityType,
-    Status,
-    UsableStatus,
-)
+from ._enums import CreatedByType
+from ._enums import KeyType
+from ._enums import ProvisioningStatus
+from ._enums import ResourceIdentityType
+from ._enums import Status
+from ._enums import UsableStatus
 from ._patch import __all__ as _patch_all
-from ._patch import *  # type: ignore # pylint: disable=unused-wildcard-import
+from ._patch import *  # pylint: disable=unused-wildcard-import
 from ._patch import patch_sdk as _patch_sdk
+
 __all__ = [
-    'CheckNameAvailabilityParameters',
-    'CheckNameAvailabilityResult',
-    'ErrorAdditionalInfo',
-    'ErrorDetail',
-    'ErrorResponse',
-    'OfferingsListResult',
-    'Operation',
-    'OperationDisplay',
-    'OperationsList',
-    'PricingDetail',
-    'PricingDimension',
-    'Provider',
-    'ProviderDescription',
-    'ProviderProperties',
-    'ProviderPropertiesAad',
-    'ProviderPropertiesManagedApplication',
-    'QuantumWorkspace',
-    'QuantumWorkspaceIdentity',
-    'QuotaDimension',
-    'Resource',
-    'SkuDescription',
-    'SystemData',
-    'TagsObject',
-    'TargetDescription',
-    'TrackedResource',
-    'WorkspaceListResult',
-    'CreatedByType',
-    'ProvisioningStatus',
-    'ResourceIdentityType',
-    'Status',
-    'UsableStatus',
+    "APIKeys",
+    "ApiKey",
+    "CheckNameAvailabilityParameters",
+    "CheckNameAvailabilityResult",
+    "ErrorAdditionalInfo",
+    "ErrorDetail",
+    "ErrorResponse",
+    "ListKeysResult",
+    "Operation",
+    "OperationDisplay",
+    "PricingDetail",
+    "PricingDimension",
+    "Provider",
+    "ProviderDescription",
+    "ProviderProperties",
+    "ProviderPropertiesAad",
+    "ProviderPropertiesManagedApplication",
+    "QuantumWorkspace",
+    "QuantumWorkspaceIdentity",
+    "QuotaDimension",
+    "Resource",
+    "SkuDescription",
+    "SystemData",
+    "TagsObject",
+    "TargetDescription",
+    "TrackedResource",
+    "WorkspaceResourceProperties",
+    "CreatedByType",
+    "KeyType",
+    "ProvisioningStatus",
+    "ResourceIdentityType",
+    "Status",
+    "UsableStatus",
 ]
 __all__.extend([p for p in _patch_all if p not in __all__])
 _patch_sdk()